Upgrade to 0.29.1
