+++ android/app/src/main/res/values/strings.xml
@@ -3,5 +3,6 @@
 
 
 
+
     <string name="app_name">karakuri</string>
 </resources>
