+++ package.json
@@ -9,7 +9,7 @@
     "apsl-react-native-button": "^2.5.0",
     "lodash": "^4.13.1",
     "react": "^15.2.1",
-    "react-native": "^0.28.0",
+    "react-native": "^0.29.1",
     "react-native-alphabetlistview": "^0.2.0",
     "react-native-drawer": "^2.2.3",
     "react-native-vector-icons": "^2.0.3"
